feat: add conditional static export config and export route fixes
- Extract createNextConfig() for testability (6 unit tests)
- NEXT_EXPORT=true enables output:export, images.unoptimized, disables PWA
- Add force-static directive to robots.ts, sitemap.ts, and route handlers for export compat
- Add generateStaticParams() wrapper pages for all dynamic [id] routes
- Make auth layouts export-compatible with conditional cookies() skip
- All existing standalone build + tests still pass (verified)
