+++ apps/web/next.config.ts
@@ -7,10 +7,11 @@ const apiBase = (
 ).replace(/\/api$/, "");
 const { protocol, hostname, port } = new URL(apiBase);
 const isDev = process.env.NODE_ENV === "development";
+const isExport = process.env.NEXT_EXPORT === "true";
 
 const withPWA = withPWAInit({
   dest: "public",
-  disable: isDev,
+  disable: isDev || isExport,
   cacheStartUrl: true,
   dynamicStartUrl: true,
   customWorkerSrc: "worker",
@@ -79,27 +80,37 @@ const withPWA = withPWAInit({
   },
 });
 
-const nextConfig: NextConfig = {
-  output: "standalone",
-  transpilePackages: ["@journiful/shared"],
-  reactStrictMode: true,
-  turbopack: {},
-  experimental: {
-    viewTransition: true,
-    optimizePackageImports: ["lucide-react", "date-fns"],
-  },
-  images: {
-    remotePatterns: [
-      {
-        protocol: protocol.replace(":", "") as "http" | "https",
-        hostname,
-        ...(port ? { port } : {}),
-        pathname: "/uploads/**",
-      },
-    ],
-    // Allow localhost/private IPs in dev (Next.js blocks them by default)
-    dangerouslyAllowLocalIP: isDev,
-  },
-};
+export function createNextConfig(
+  env: Record<string, string | undefined>,
+): NextConfig {
+  const isExportMode = env.NEXT_EXPORT === "true";
+
+  return {
+    output: isExportMode ? "export" : "standalone",
+    transpilePackages: ["@journiful/shared"],
+    reactStrictMode: true,
+    turbopack: {},
+    experimental: {
+      viewTransition: true,
+      optimizePackageImports: ["lucide-react", "date-fns"],
+    },
+    images: isExportMode
+      ? { unoptimized: true }
+      : {
+          remotePatterns: [
+            {
+              protocol: protocol.replace(":", "") as "http" | "https",
+              hostname,
+              ...(port ? { port } : {}),
+              pathname: "/uploads/**",
+            },
+          ],
+          // Allow localhost/private IPs in dev (Next.js blocks them by default)
+          dangerouslyAllowLocalIP: isDev,
+        },
+  };
+}
+
+const nextConfig: NextConfig = createNextConfig(process.env);
 
 export default withPWA(nextConfig);
