+++ apps/web/src/app/(app)/trips/[id]/itinerary/tab-page.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { useTripPage } from "../trip-page-context";
+import { ItineraryView } from "@/components/itinerary/itinerary-view";
+
+export default function ItineraryTab() {
+  const { tripId, weather, temperatureUnit, setShowOnboarding } = useTripPage();
+
+  return (
+    <ItineraryView
+      tripId={tripId}
+      onAddTravel={() => setShowOnboarding(true)}
+      forecasts={weather?.forecasts}
+      temperatureUnit={temperatureUnit}
+    />
+  );
+}
