+++ apps/web/src/app/(app)/trips/[id]/page.tsx
@@ -1,17 +1,14 @@
-"use client";
+import TripDefaultPage from "./default-tab-page";
 
-import { useTripPage } from "./trip-page-context";
-import { ItineraryView } from "@/components/itinerary/itinerary-view";
+export function generateStaticParams() {
+  return [{ id: "static-export-placeholder" }];
+}
 
-export default function TripDefaultPage() {
-  const { tripId, weather, temperatureUnit, setShowOnboarding } = useTripPage();
-
-  return (
-    <ItineraryView
-      tripId={tripId}
-      onAddTravel={() => setShowOnboarding(true)}
-      forecasts={weather?.forecasts}
-      temperatureUnit={temperatureUnit}
-    />
-  );
+export default function TripDefaultPageWrapper({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  void params;
+  return <TripDefaultPage />;
 }
