+++ apps/web/src/app/(auth)/login/layout.tsx
@@ -3,6 +3,8 @@ import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import type { ReactNode } from "react";
 
+export const dynamic = "force-static";
+
 export const metadata: Metadata = {
   title: "Sign In",
   description:
@@ -15,11 +17,16 @@ export default async function LoginLayout({
 }: {
   children: ReactNode;
 }) {
-  const cookieStore = await cookies();
-  const authToken = cookieStore.get("auth_token");
+  // In static export mode, cookies are unavailable — skip auth check
+  const isExport = process.env.NEXT_EXPORT === "true";
 
-  if (authToken?.value) {
-    redirect("/trips");
+  if (!isExport) {
+    const cookieStore = await cookies();
+    const authToken = cookieStore.get("auth_token");
+
+    if (authToken?.value) {
+      redirect("/trips");
+    }
   }
 
   return children;
