+++ apps/web/src/app/twitter-image.tsx
@@ -1,1 +1,2 @@
 export { default, alt, size, contentType } from "./opengraph-image";
+export const dynamic = "force-static";
